+++ packages/react/modules/bookmark/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 5.0.2-cli-search-index.0
+
+### Patch Changes
+
+- [#3757](https://github.com/equinor/fusion-framework/pull/3757) [`db880d1`](https://github.com/equinor/fusion-framework/commit/db880d1fbdb62ba4667f11229d1e6c3a4cea06fc) Thanks [@odinr](https://github.com/odinr)! - preview release
+
+- Updated dependencies [[`db880d1`](https://github.com/equinor/fusion-framework/commit/db880d1fbdb62ba4667f11229d1e6c3a4cea06fc)]:
+  - @equinor/fusion-framework-module-bookmark@3.0.4-cli-search-index.0
+  - @equinor/fusion-framework-module@5.0.6-cli-search-index.0
+  - @equinor/fusion-framework-react@7.4.20-cli-search-index.0
+  - @equinor/fusion-framework-react-module@3.1.14-cli-search-index.0
+  - @equinor/fusion-observable@8.5.7-cli-search-index.0
+
 ## 5.0.1
 
 ### Patch Changes