+++ packages/widget/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.1.47-cli-search-index.0
+
+### Patch Changes
+
+- [#3757](https://github.com/equinor/fusion-framework/pull/3757) [`db880d1`](https://github.com/equinor/fusion-framework/commit/db880d1fbdb62ba4667f11229d1e6c3a4cea06fc) Thanks [@odinr](https://github.com/odinr)! - preview release
+
+- Updated dependencies [[`db880d1`](https://github.com/equinor/fusion-framework/commit/db880d1fbdb62ba4667f11229d1e6c3a4cea06fc)]:
+  - @equinor/fusion-framework@7.4.3-cli-search-index.0
+  - @equinor/fusion-framework-module-event@4.4.1-cli-search-index.0
+  - @equinor/fusion-framework-module-http@7.0.6-cli-search-index.0
+  - @equinor/fusion-framework-module@5.0.6-cli-search-index.0
+  - @equinor/fusion-framework-module-msal@6.0.1-cli-search-index.0
+  - @equinor/fusion-framework-module-widget@12.0.3-cli-search-index.0
+
 ## 1.1.46
 
 ### Patch Changes